+++ logscollector/collector.go
@@ -15,22 +15,24 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
-type LogsCollector struct {
+const timeFormat = time.RFC3339
+
+type Collector struct {
 	logsDir   string
 	outputDir string
 
 	filesToCollect chan string
 }
 
-func New(logsDir, outputDir string) *LogsCollector {
-	return &LogsCollector{
+func New(logsDir, outputDir string) *Collector {
+	return &Collector{
 		filesToCollect: make(chan string, 20),
 		logsDir:        logsDir,
 		outputDir:      outputDir,
 	}
 }
 
-func (lc LogsCollector) watchForCreatedFiles(ctx context.Context) error {
+func (lc Collector) watchForCreatedFiles(ctx context.Context) error {
 	watcher, err := fsnotify.NewWatcher()
 	if err != nil {
 		return err
@@ -65,7 +67,7 @@ func (lc LogsCollector) watchForCreatedFiles(ctx context.Context) error {
 	}
 }
 
-func (lc LogsCollector) collectLogs(ctx context.Context, logFilePath string) error {
+func (lc Collector) collectLogs(ctx context.Context, logFilePath string) error {
 	log.Printf("Setting up file listener for %q", logFilePath)
 	defer log.Printf("Shutting down file listener for %q", logFilePath)
 
@@ -77,7 +79,7 @@ func (lc LogsCollector) collectLogs(ctx context.Context, logFilePath string) err
 
 	logFileNameBase := path.Base(logFilePath)
 	logFileName := strings.TrimSuffix(logFileNameBase, filepath.Ext(logFileNameBase))
-	destLogFile := path.Join(lc.outputDir, fmt.Sprintf("%s-%s.log", logFileName, time.Now().Format(time.RFC3339)))
+	destLogFile := path.Join(lc.outputDir, fmt.Sprintf("%s-%s.log", logFileName, time.Now().Format(timeFormat)))
 
 	f, err := os.Create(destLogFile)
 	if err != nil {
@@ -118,7 +120,7 @@ func (lc LogsCollector) collectLogs(ctx context.Context, logFilePath string) err
 	}
 }
 
-func (lc LogsCollector) Run(ctx context.Context) error {
+func (lc Collector) Run(ctx context.Context) error {
 	match := fmt.Sprintf("%s/*.[stderr|stdout]*[0-9]", lc.logsDir)
 
 	log.Printf("Looking for log files with %q", match)